refactor(web): split github-core/mutations.ts into resource sub-modules behind a barrel
Decompose the 2842-line writes monolith into cohesive ./mutations/* sub-modules
(gitObjects leaf, teams, orgMembership, provisioning, secrets, workflowDispatch,
collaborators, classroomEdit). gitObjects.ts is the shared git-data leaf
(tree/commit/ref/blob + classroom-seed helpers) that provisioning + classroomEdit
import downward; it takes a type-only edge to ./teams for the ref types so the
no-cycle guard stays green. provisioning.ts keeps the initClassroom50 orchestrator
+ setup steps together (~1306 lines, the acceptable orchestrator anchor).

mutations.ts is now a thin re-export barrel: all 83 exported names preserved
(incl. isNonFastForward, which domain/classrooms.ts depends on), so all ~31
importers are unchanged. Behavior-preserving — tsc + full vitest (1533) green,
no-cycle + boundaries + arch:validate green; mutations.test.ts +
mutations.deleteTeam.test.ts still exercise the public surface via the barrel.
